verso
Struct VersoBuilder

pub struct VersoBuilder(/* private fields */);

A builder for configuring and creating a VersoviewController instance.

Implementations

impl VersoBuilder

pub fn new() -> Self

Creates a new VersoBuilder with default settings.

pub fn with_panel(self, with_panel: bool) -> Self

Sets whether the control panel should be included.

pub fn inner_size(self, size: impl Into<Size>) -> Self

Sets the initial window size.

pub fn position(self, position: impl Into<Position>) -> Self

Sets the initial window position.

pub fn maximized(self, maximized: bool) -> Self

Sets whether the window should start maximized.

pub fn visible(self, visible: bool) -> Self

Sets whether the window should be visible initially.

pub fn fullscreen(self, fullscreen: bool) -> Self

Sets whether the window should start in fullscreen mode.

pub fn focused(self, focused: bool) -> Self

Sets whether the window will be initially focused or not.

pub fn devtools_port(self, port: u16) -> Self

Port number to start a server to listen to remote Firefox devtools connections. 0 for random port.

pub fn profiler_settings(self, settings: ProfilerSettings) -> Self

Sets the profiler settings.

pub fn user_agent(self, user_agent: impl Into<String>) -> Self

Overrides the user agent.

pub fn userscripts_directory(self, directory: impl Into<String>) -> Self

Sets the script to run when the document starts loading. Sets the directory to load user scripts from.

pub fn zoom_level(self, zoom: f32) -> Self

Sets the initial zoom level of the webview.

pub fn resources_directory(self, path: impl Into<PathBuf>) -> Self

Sets the resource directory path.

pub fn build( self, versoview_path: impl AsRef<Path>, initial_url: Url, ) -> VersoviewController

Builds the VersoviewController with the configured settings.