verso
Struct VersoviewController

pub struct VersoviewController { /* private fields */ }

A VersoView controller

Send an exit signal to this Verso instance when dropped

Implementations

impl VersoviewController

pub fn new(verso_path: impl AsRef<Path>, initial_url: Url) -> Self

Create a new verso instance with default settings and get the controller to it

pub fn exit(&self) -> Result<(), Box<ErrorKind>>

Exit

pub fn on_close_requested( &self, callback: impl Fn() + Send + 'static, ) -> Result<(), Box<ErrorKind>>

Listen on close requested from the OS, if you decide to use it, verso will not close the window by itself anymore, so make sure you handle it properly by either do your own logic or call Self::exit as a fallback

pub fn execute_script(&self, script: String) -> Result<(), Box<ErrorKind>>

Execute script

pub fn navigate(&self, url: Url) -> Result<(), Box<ErrorKind>>

Navigate to url

pub fn on_navigation_starting( &self, callback: impl Fn(Url) -> bool + Send + 'static, ) -> Result<(), Box<ErrorKind>>

Listen on navigation starting triggered by user click on a link, return a boolean in the callback to decide whether or not allowing this navigation

pub fn on_web_resource_requested( &self, callback: impl Fn(WebResourceRequest, Box<dyn FnOnce(Option<Response<Vec<u8>>>) + Send>) + Send + 'static, ) -> Result<(), Box<ErrorKind>>

Listen on web resource requests, return a boolean in the callback to decide whether or not allowing this navigation

pub fn set_size<S: Into<Size>>(&self, size: S) -> Result<(), Box<ErrorKind>>

Sets the webview window’s size

pub fn set_position<P: Into<Position>>( &self, position: P, ) -> Result<(), Box<ErrorKind>>

Sets the webview window’s position

pub fn set_maximized(&self, maximized: bool) -> Result<(), Box<ErrorKind>>

Maximize or unmaximize the window

pub fn set_minimized(&self, minimized: bool) -> Result<(), Box<ErrorKind>>

Minimize or unminimize the window

pub fn set_fullscreen(&self, fullscreen: bool) -> Result<(), Box<ErrorKind>>

Sets the window to fullscreen or back

pub fn set_visible(&self, visible: bool) -> Result<(), Box<ErrorKind>>

Show or hide the window

pub fn start_dragging(&self) -> Result<(), Box<ErrorKind>>

Moves the window with the left mouse button until the button is released

pub fn get_inner_size(&self) -> Result<PhysicalSize<u32>, Box<ErrorKind>>

Returns the physical size of the window’s client area.

The client area is the content of the window, excluding the title bar and borders.

pub fn get_outer_size(&self) -> Result<PhysicalSize<u32>, Box<ErrorKind>>

Returns the physical size of the entire window.

These dimensions include the title bar and borders. If you don’t want that (and you usually don’t), use Self::get_inner_size instead.

pub fn get_inner_position( &self, ) -> Result<Option<PhysicalPosition<i32>>, Box<ErrorKind>>

Get the window’s inner position, returns None on unsupported platforms (currently only Wayland)

pub fn get_outer_position( &self, ) -> Result<Option<PhysicalPosition<i32>>, Box<ErrorKind>>

Get the window’s outer position, returns None on unsupported platforms (currently only Wayland)

pub fn is_maximized(&self) -> Result<bool, Box<ErrorKind>>

Get if the window is currently maximized or not

pub fn is_minimized(&self) -> Result<bool, Box<ErrorKind>>

Get if the window is currently minimized or not

pub fn is_fullscreen(&self) -> Result<bool, Box<ErrorKind>>

Get if the window is currently fullscreen or not

pub fn is_visible(&self) -> Result<bool, Box<ErrorKind>>

Get the visibility of the window

pub fn get_scale_factor(&self) -> Result<f64, Box<ErrorKind>>

Get the scale factor of the window

pub fn get_current_url(&self) -> Result<Url, Box<ErrorKind>>

Get the URL of the webview